t3dlib1 : remove TEXTUREV1
t3dlib5 : enhance RENDERCONTEXT

t3dlib6 : bug fix CAM4DV1::viewport::x, y type : int -> REAL

git-svn-id: svn://127.0.0.1/T3DLIB2/trunk@38 6c76846e-3230-0410-94b0-57f3116ac674

+++ T3DLIB2/include/t3dlib6.h
@@ -126,7 +126,7 @@ typedef struct T3DLIB_API CAM4DV1_TYP
 
 	struct VIEWPORT_TYP
 	{
-		int			x, y;
+		REAL		x, y;
 		REAL		width, height;
 
 	} viewport;
